+++ x/btcstaking/keeper/genesis_test.go
@@ -46,8 +46,6 @@ func TestExportGenesis(t *testing.T) {
 	fps := datagen.CreateNFinalityProviders(r, t, numFps)
 	params := k.GetParams(ctx)
 
-	minUnbondingTime := params.MinUnbondingTimeBlocks
-
 	chainsHeight := make([]*types.BlockHeightBbnToBtc, 0)
 	// creates the first as it starts already with an chain height from the helper.
 	chainsHeight = append(chainsHeight, &types.BlockHeightBbnToBtc{
@@ -86,7 +84,7 @@ func TestExportGenesis(t *testing.T) {
 			totalDelegations++
 
 			// sets delegations
-			h.AddDelegation(del, minUnbondingTime)
+			h.AddDelegation(del)
 			btcDelegations = append(btcDelegations, del)
 
 			// BTC delegators idx
@@ -115,7 +113,7 @@ func TestExportGenesis(t *testing.T) {
 			idxEvent := uint64(totalDelegations - 1)
 			eventsIdx[idxEvent] = &types.EventIndex{
 				Idx:            idxEvent,
-				BlockHeightBtc: del.EndHeight - minUnbondingTime,
+				BlockHeightBtc: del.EndHeight - del.UnbondingTime,
 				Event:          unbondedEvent,
 			}
 		}