+++ x/btcstaking/keeper/msg_server_test.go
@@ -48,7 +48,7 @@ func FuzzMsgServer_UpdateParams(f *testing.F) {
 
 		params := h.BTCStakingKeeper.GetParams(h.Ctx)
 		ckptFinalizationTimeout := btccKeeper.GetParams(h.Ctx).CheckpointFinalizationTimeout
-		params.MinUnbondingTimeBlocks = uint32(r.Intn(int(ckptFinalizationTimeout))) + 1
+		params.UnbondingTimeBlocks = uint32(r.Intn(int(ckptFinalizationTimeout))) + 1
 		params.BtcActivationHeight = params.BtcActivationHeight + 1
 
 		// Try to update params with minUnbondingTime less than or equal to checkpointFinalizationTimeout
@@ -62,7 +62,7 @@ func FuzzMsgServer_UpdateParams(f *testing.F) {
 			"should not set minUnbondingTime to be less than checkpointFinalizationTimeout")
 
 		// Try to update params with minUnbondingTime larger than checkpointFinalizationTimeout
-		msg.Params.MinUnbondingTimeBlocks = uint32(r.Intn(1000)) + ckptFinalizationTimeout + 1
+		msg.Params.UnbondingTimeBlocks = uint32(r.Intn(1000)) + ckptFinalizationTimeout + 1
 		_, err = h.MsgServer.UpdateParams(h.Ctx, msg)
 		require.NoError(t, err)
 	})
@@ -360,7 +360,7 @@ func TestProperVersionInDelegation(t *testing.T) {
 
 	customMinUnbondingTime := uint32(2000)
 	currentParams := h.BTCStakingKeeper.GetParams(h.Ctx)
-	currentParams.MinUnbondingTimeBlocks = 2000
+	currentParams.UnbondingTimeBlocks = customMinUnbondingTime
 	currentParams.BtcActivationHeight = currentParams.BtcActivationHeight + 1
 	// Update new params
 	err = h.BTCStakingKeeper.SetParams(h.Ctx, currentParams)
@@ -374,7 +374,7 @@ func TestProperVersionInDelegation(t *testing.T) {
 		stakingValue,
 		10000,
 		stakingValue-1000,
-		uint16(customMinUnbondingTime)+1,
+		uint16(customMinUnbondingTime),
 		false,
 		false,
 	)
@@ -802,9 +802,9 @@ func TestDoNotAllowDelegationWithoutFinalityProvider(t *testing.T) {
 	_, covenantPKs := h.GenAndApplyParams(r)
 	bsParams := h.BTCStakingKeeper.GetParams(h.Ctx)
 
-	minUnbondingTime := bsParams.MinUnbondingTimeBlocks
+	unbondingTime := bsParams.UnbondingTimeBlocks
 
-	slashingChangeLockTime := uint16(minUnbondingTime)
+	slashingChangeLockTime := uint16(unbondingTime)
 
 	// We only generate a finality provider, but not insert it into KVStore. So later
 	// insertion of delegation should fail.
@@ -866,7 +866,6 @@ func TestDoNotAllowDelegationWithoutFinalityProvider(t *testing.T) {
 	require.NoError(t, err)
 
 	stkTxHash := testStakingInfo.StakingTx.TxHash()
-	unbondingTime := bsParams.MinUnbondingTimeBlocks
 	unbondingValue := stakingValue - datagen.UnbondingTxFee // TODO: parameterise fee
 	testUnbondingInfo := datagen.GenBTCUnbondingSlashingInfo(
 		r,
@@ -929,30 +928,23 @@ func TestCorrectUnbondingTimeInDelegation(t *testing.T) {
 	tests := []struct {
 		name                      string
 		finalizationTimeout       uint32
-		minUnbondingTime          uint32
+		unbondingTimeInParams     uint32
 		unbondingTimeInDelegation uint16
 		err                       error
 	}{
 		{
-			name:                      "successful delegation when ubonding time in delegation is larger than finalization timeout when finalization timeout is larger than min unbonding time",
+			name:                      "successful delegation if unbonding time in delegation is equal to unbonding time in params",
 			unbondingTimeInDelegation: 101,
-			minUnbondingTime:          99,
+			unbondingTimeInParams:     101,
 			finalizationTimeout:       100,
 			err:                       nil,
 		},
 		{
-			name:                      "successful delegation when ubonding time ubonding time in delegation is larger than min unbonding time when min unbonding time is larger than finalization timeout",
-			unbondingTimeInDelegation: 151,
-			minUnbondingTime:          150,
+			name:                      "invalid delegation if unbonding time is different from unbonding time in params",
+			unbondingTimeInDelegation: 102,
+			unbondingTimeInParams:     101,
 			finalizationTimeout:       100,
-			err:                       nil,
-		},
-		{
-			name:                      "successful delegation when ubonding time in delegation is equal to minUnbondingTime when min unbonding time is larger than finalization timeout",
-			unbondingTimeInDelegation: 150,
-			minUnbondingTime:          150,
-			finalizationTimeout:       100,
-			err:                       nil,
+			err:                       types.ErrInvalidUnbondingTx,
 		},
 	}
 
@@ -968,7 +960,7 @@ func TestCorrectUnbondingTimeInDelegation(t *testing.T) {
 			h := testutil.NewHelper(t, btclcKeeper, btccKeeper)
 
 			// set all parameters
-			_, _ = h.GenAndApplyCustomParams(r, tt.finalizationTimeout, tt.minUnbondingTime, 0)
+			_, _ = h.GenAndApplyCustomParams(r, tt.finalizationTimeout, tt.unbondingTimeInParams, 0)
 
 			changeAddress, err := datagen.GenRandomBTCAddress(r, h.Net)
 			require.NoError(t, err)
@@ -1177,7 +1169,6 @@ func FuzzDeterminismBtcstakingBeginBlocker(f *testing.F) {
 		covQuorum := stakingParams.CovenantQuorum
 		maxFinalityProviders := int32(h.App.FinalityKeeper.GetParams(h.Ctx).MaxActiveFinalityProviders)
 
-		minUnbondingTime := stakingParams.MinUnbondingTimeBlocks
 		// Number of finality providers from 10 to maxFinalityProviders + 10
 		numFinalityProviders := int(r.Int31n(maxFinalityProviders) + 10)
 
@@ -1211,8 +1202,8 @@ func FuzzDeterminismBtcstakingBeginBlocker(f *testing.F) {
 			)
 
 			for _, del := range delegations {
-				h.AddDelegation(del, minUnbondingTime)
-				h1.AddDelegation(del, minUnbondingTime)
+				h.AddDelegation(del)
+				h1.AddDelegation(del)
 			}
 		}